refactor gatsby-node page generation
trim unnecessary fields off query.
get all data from single query, rather than pulling from two different queries.

the way it is currently set up, the assumption is that ALL bikes have a slug field and not a path field,
while ALL posts have a path field but not a slug field. I'm sure that this structure could be improved,
but for now it works as needed.